requirements.txt
tensorflow
opencv-python
numpy
fastapi
uvicorn[standard]
python-multipart
Pillow
